B: Making util.cv module
