+++ opsi/modules/mask.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass
 
-import opsi.manager.cvwrapper as cvw
+import opsi.util.cv._wrappers as cvw
 from opsi.manager.manager_schema import Function
 from opsi.manager.types import MatBW
