+++ opsi/util/cv/__init__.py
@@ -1,0 +1,3 @@
+from .contour import Contour, Contours
+from .mat import Mat, MatBW
+from .shape import Point, Rect
